refactor(projects/art-jam-i): Move portrait under components/p5; add support for vite to recognize  paths to avoid relative imports
