+++ projects/art-jam-i/package.json
@@ -36,6 +36,7 @@
     "globals": "^16.3.0",
     "jest": "^30.1.3",
     "jest-environment-jsdom": "^30.1.2",
+    "path": "^0.12.7",
     "prettier": "^3.6.2",
     "vite": "^7.1.2"
   }
