+++ projects/art-jam-i/src/components/p5/portrait.js
@@ -6,9 +6,9 @@
  * This is just a P5/canvas-friendly wrapper for the portrait.svg file
  */
 
-import SVGDrawer from "../utils/svg";
+import SVGDrawer from "src/utils/svg";
 // `?raw` tells vite to return the import as a string rather than wrap it as a React component
-import portraitSvgRaw from "./portrait.svg?raw";
+import portraitSvgRaw from "src/assets/portrait.svg?raw";
 // import portraitGlassesSvgRaw from './portrait-glasses.svg?raw';
 
 export default class Portrait {
